Moved aoi atoms to own file
