+++ app/scripts/components/common/map/controls/hooks/use-aois.ts
@@ -2,12 +2,7 @@ import { useAtomValue, useSetAtom } from 'jotai';
 import { useCallback } from 'react';
 import { Polygon } from 'geojson';
 import { toAoIid } from '../../utils';
-import {
-  aoisDeleteAtom,
-  aoisFeaturesAtom,
-  aoisSetSelectedAtom,
-  aoisUpdateGeometryAtom
-} from '$components/exploration/atoms/atoms';
+import { aoisDeleteAtom, aoisFeaturesAtom, aoisSetSelectedAtom, aoisUpdateGeometryAtom } from '../aoi/atoms';
 
 export default function useAois() {
   const features = useAtomValue(aoisFeaturesAtom);
