+++ app/scripts/components/exploration/atoms/atoms.ts
@@ -1,6 +1,4 @@
 import { atom } from 'jotai';
-import { atomWithHash } from 'jotai-location';
-import { Polygon } from 'geojson';
 import {
   DataMetric,
   dataMetrics
@@ -8,8 +6,6 @@ import {
 
 import { HEADER_COLUMN_WIDTH, RIGHT_AXIS_SPACE } from '../constants';
 import { DateRange, TimelineDataset, ZoomTransformPlain } from '../types.d.ts';
-import { decodeAois, encodeAois } from '$utils/polygon-url';
-import { AoIFeature } from '$components/common/map/types';
 
 // Datasets to show on the timeline and their settings
 export const timelineDatasetsAtom = atom<TimelineDataset[]>([]);
@@ -47,55 +43,3 @@ export const activeAnalysisMetricsAtom = atom<DataMetric[]>(dataMetrics);
 
 // 🛑 Whether or not an analysis is being performed. Temporary!!!
 export const isAnalysisAtom = atom<boolean>(false);
-
-// This is the atom acting as a single source of truth for the AOIs.
-export const aoisHashAtom = atomWithHash('aois', '');
-
-// Getter atom to get AoiS as GeoJSON features from the hash.
-export const aoisFeaturesAtom = atom<AoIFeature[]>((get) => {
-  const hash = get(aoisHashAtom);
-  if (!hash) return [];
-  return decodeAois(hash);
-});
-
-// Setter atom to update AOIs geoometries, writing directly to the hash atom.
-export const aoisUpdateGeometryAtom = atom(
-  null,
-  (get, set, updates: { id: string; geometry: Polygon }[]) => {
-    let newFeatures = [...get(aoisFeaturesAtom)];
-    updates.forEach(({ id, geometry }) => {
-      const existingFeature = newFeatures.find((feature) => feature.id === id);
-      if (existingFeature) {
-        existingFeature.geometry = geometry;
-      } else {
-        const newFeature: AoIFeature = {
-          type: 'Feature',
-          id,
-          geometry,
-          selected: true,
-          properties: {}
-        };
-        newFeatures = [...newFeatures, newFeature];
-      }
-    });
-    set(aoisHashAtom, encodeAois(newFeatures));
-  }
-);
-
-// Setter atom to update AOIs selected state, writing directly to the hash atom.
-export const aoisSetSelectedAtom = atom(null, (get, set, ids: string[]) => {
-  const features = get(aoisFeaturesAtom);
-  const newFeatures = features.map((feature) => {
-    return { ...feature, selected: ids.includes(feature.id as string) };
-  });
-  set(aoisHashAtom, encodeAois(newFeatures));
-});
-
-// Setter atom to delete AOIs, writing directly to the hash atom.
-export const aoisDeleteAtom = atom(null, (get, set, ids: string[]) => {
-  const features = get(aoisFeaturesAtom);
-  const newFeatures = features.filter(
-    (feature) => !ids.includes(feature.id as string)
-  );
-  set(aoisHashAtom, encodeAois(newFeatures));
-});
